* main: (26 commits)
  chore: test with prepared statements off for pg (#4233)
  fix(storage): map protobuf enums to ints in sql updates for rule and (#4231)
  fix(audit): keep segment data for rollout and rule (#4229)
  chore(deps): bump codecov/codecov-action from 5.4.2 to 5.4.3 (#4214)
  chore(deps): bump github.com/open-policy-agent/opa from 1.4.0 to 1.4.2 (#4216)
  chore(deps): bump github.com/xo/dburl from 0.23.4 to 0.23.7 (#4217)
  chore(deps): bump go.opentelemetry.io/contrib/instrumentation/runtime (#4218)
  chore(deps): bump github.com/aws/aws-sdk-go-v2/service/s3 (#4219)
  chore(deps): bump @loadable/component from 5.16.4 to 5.16.7 in /ui (#4224)
  chore(deps-dev): bump vite from 6.3.4 to 6.3.5 in /ui (#4220)
  chore(deps): bump @tanstack/react-table from 8.21.2 to 8.21.3 in /ui (#4223)
  chore(deps): bump chart.js from 4.4.8 to 4.4.9 in /ui (#4222)
  chore(deps): bump @radix-ui/react-select from 2.1.7 to 2.2.4 in /ui (#4221)
  chore(deps): bump github.com/golang-migrate/migrate/v4 (#4215)
  chore(deps): bump golang.org/x/tools from 0.32.0 to 0.33.0 in /_tools (#4199)
  chore(deps): bump golang.org/x/crypto from 0.37.0 to 0.38.0 (#4192)
  chore(deps): bump github.com/go-sql-driver/mysql from 1.9.1 to 1.9.2 (#4189)
  chore(deps): bump cloud.google.com/go/storage from 1.51.0 to 1.53.0 (#4193)
  chore(deps): bump zod from 3.24.3 to 3.24.4 in /ui (#4195)
  chore(deps): bump react-router from 7.5.3 to 7.6.0 in /ui (#4196)
  ...

+++ internal/storage/sql/common/rollout.go
@@ -585,7 +585,7 @@ func (s *Store) UpdateRollout(ctx context.Context, r *flipt.UpdateRolloutRequest
 
 		if _, err := s.builder.Update(tableRolloutSegments).
 			RunWith(tx).
-			Set("segment_operator", segmentOperator).
+			Set("segment_operator", int32(segmentOperator)).
 			Set("value", segmentRule.Value).
 			Where(sq.Eq{"rollout_id": r.Id}).ExecContext(ctx); err != nil {
 			return nil, err
